requirements.txt
Django==2.0.5
djangorestframework==3.8.2
pytz==2018.4
django-filter
